+++ src/main/resources/templates/board/detail.html
@@ -6,7 +6,7 @@
 <script src="/js/board.js" defer="defer"></script>
 </head>
 <body>
-	<header th:include="/headerNav"></header>
+	<header th:include="/nav/headerNav"></header>
 	<main class="container mt-5">
 		<div>
 			<p>
