+++ src/main/resources/templates/board/list.html
@@ -25,27 +25,101 @@
 	height:400px;
 	border:1px solid;
 }
+
+.modalbutton button{
+width:250px;
+}
 </style>
 </head>
 <body>
-<header th:include="/headerNav"></header>
-<a th:href="@{/board/insert.do}">글쓰기</a>
-<main class="container ">
-	<div class="board"
-	th:onclick="location.href='detail.do?boardNo=[[${board.boardNo}]]'"
-	th:each="board:${boardList}">
-		<div>
-			<div class="mainBoardImg row row-cols-2">
-				<th:block th:each="img:${board.boardImgList}" >
-					<img class="col" style="width: 100px; height: 100px;" th:src="@{/img/board/{file}(file=${img.imgPath})}">
-				</th:block>
+		<div class="board"
+		th:onclick="location.href='detail.do?boardNo=[[${board.boardNo}]]'"
+		th:each="board:${boardList}">
+			<div>
+				<div class="mainBoardImg row row-cols-2">
+					<th:block th:each="img:${board.boardImgList}" >
+						<img class="col" style="width: 100px; height: 100px;" th:src="@{/img/board/{file}(file=${img.imgPath})}">
+					</th:block>
+				</div>
+				<div class="mainBoardContent row">
+					<span th:text="${board.contents}"></span>
+				</div>
 			</div>
-			<div class="mainBoardContent row">
-				<span th:text="${board.contents}"></span>
-			</div>
+			<div class="mainBoardReply"></div>
 		</div>
-		<div class="mainBoardReply"></div>
-	</div>
-</main>
+		
+<!-- 		<a class="btn btn-outline-success" th:href="@{/board/insert.do}">글쓰기</a> -->
+		<div class="d-flex justify-content-end">
+			<button type="button"
+			class="btn btn-outline-primary m-5 boardInBtn" data-bs-toggle="modal" data-bs-target="#exampleModal">
+			  글쓰기
+			</button>
+		</div>
+		
+		<!-- Modal -->
+		<div class="modal fade" id="exampleModal" tabindex="-1" aria-labelledby="exampleModalLabel" aria-hidden="true">
+		  <div class="modal-dialog modal-dialog-centered modal-xl">
+		    <div class="modal-content">
+  		        <form  enctype="multipart/form-data" action="/board/insert.do" method="POST">
+  		        
+		      <div class="modal-header">
+		        <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
+		      </div>
+		      
+		      <div class="modal-body">
+				<input type="hidden" value="1" name="userNo">
+
+				<div class="d-flex">
+					<div class="boardInputLeft">
+						<div class="input-group mb-3">
+							 <label class="input-group-text" for="selCategory">Choose</label>
+							 <select class="form-select" id="selCategory" name="category">
+								   <option selected>Category</option>
+								   <option value="food">Food</option>
+								   <option value="sport">Sport</option>
+								   <option value="sights">Sights</option>
+								   <option value="etc">etc</option>
+							 </select>
+						</div>
+						
+						<div class="d-flex">
+							<div class="input-group mb-3" id="uploadBox">
+						 		<input name="img" class="form-control" type="file" id="inputImg" multiple onchange="setThumbnail(event)">
+							</div>
+							<div class="input-group mb-3 ms-3 ">
+								 <input type="text" class="form-control" placeholder="HashTag(#)">
+							</div>
+						</div>
+						
+						<div class="input-group mb-3">
+							 <label class="input-group-text" for="map">지도</label>
+							 <input type="number" class="form-control" id="map" value=0>
+						</div>
+												
+						
+						<div class="form-floating">
+							<textarea style="resize: none; height: 300px;"
+								class="form-control" name="contents" id="boardContents" placeholder="내용"></textarea>
+							<label for="boardContents">내용</label>
+						</div>
+						
+					</div>
+					
+					<div class="boarderInputRights d-flex justify-content-center">
+						<div id="imgContainer"></div>
+					</div>
+				</div>
+		      </div>
+		      
+		      <div class="modal-footer d-flex justify-content-center modalbutton">
+				<button class="btn btn-outline-primary me-4">글쓰기</button>
+				<button class="btn btn-outline-danger me-4" type="reset">초기화</button>
+		        <button type="button" class="btn btn-outline-secondary" data-bs-dismiss="modal">취소</button>
+		      </div>
+		      
+		      </form>
+		    </div>
+		  </div>
+		</div>
 </body>
 </html>
