+++ src/main/resources/templates/index.html
@@ -15,23 +15,7 @@ border : 1px solid green;
 		<main class="d-flex w-100 h-100">
 			<div class="sideNav" style="width:170px" th:include="/nav/asideNav"></div>
 			<div class="container justify-content-center bg-light ">
-			 	<div class="mt-5">
-          <div>
-            <a  class="btn"  href="/user/signup.do">회원가입 페이지로 이동</a>
-            <a  class="btn"  href="/user/login.do">로그인 페이지로 이동</a>
-            <a  class="btn"  href="/login">test</a>
-            <a  class="btn"  href="/">Home</a>
-            <span>test
-	        	<a href="/board/list.do">boardListMaintest</a>
-	        	</span>
-          </div>
-		 			<h2>힘들다 제발 진짜</h2>
-		 			<ol>
-		 				<li>
-		 					<a href="/reply/list.do">아 댓글 제발</a>
-		 				</li>
-		 			</ol>
-		 		</div>
+			 	<section class="mt-5" th:include="/board/list"></section>
 			</div>
 		</main>
 	</body>
